requirements.txt
streamlit
python-dotenv
mcp
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-ollama==0.1.3
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic